Optionally allow caller to provide minimatch options - Fix #6
If provided, the specified minimatch options will be passed to
gulp-ignore and ultimately passed to minimatch itself.

This feature is required for a specific use case: when the caller's glob
includes dot files (e.g. gulp.src('**/*.json', {dot: true})).  However,
the ability to specify the minimatch options may be useful in other
contexts, as well.